backend/requirements.txt
fastapi~=0.115.0
uvicorn[standard]~=0.30.0
pandas~=2.2.0
numpy~=1.26.0
akshare~=1.14.0
baostock~=0.8.0
pyarrow~=14.0
pyyaml~=6.0
loguru~=0.7.0
pydantic~=2.9.0
optuna~=3.6.0
